pyproject.toml
[tool.poetry]
name = "project-fibonacci"
version = "0.1.0"
description = ""
authors = ["raymonterola <83757268+raymonterola@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "project_fibonacci"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
flask-marshmallow = "^0.14.0"
celery = "^5.2.7"
redis = "^4.4.0"
flask-sqlalchemy = "^3.0.2"
marshmallow-sqlalchemy = "^0.28.1"
psycopg2-binary = "^2.9.5"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.0"
pytest = "^7.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
